refactor(runner): split dispatchFlow and the json failure collector out of run
runInternals.ts and run.ts were both at the 150-line cap, so neither could
take another field or branch. runInternals.ts now holds only types.
